연산자

+++ 221226/hw1.c
@@ -3,10 +3,10 @@
 
 int main() {
 
-    int r;
+    double r;
     printf("원의 반지름을 입력하세요.\n반지름 : ");
-    scanf("%d", &r);
-    printf("원의 넓이는 %.2f입니다.\n", r * r * PI);
+    scanf("%lf", &r);
+    printf("원의 넓이는 %.2lf입니다.\n", r * r * PI);
 
     return 0;
 }
